+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- Added torchbearer.variational, a sub-package for implementations of state of the art variational auto-encoders
 ### Changed
 ### Deprecated
 ### Removed